requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohappyeyeballs==2.6.1
aiohttp==3.8.6
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.2
frozenlist==1.7.0
greenlet==3.2.4
idna==3.10
magic-filter==1.0.12
multidict==6.6.3
propcache==0.3.2
pydantic==1.10.22
pydantic_core==2.33.2
python-dotenv==1.1.1
SQLAlchemy==2.0.42
typing-inspection==0.4.1
typing_extensions==4.14.1
yarl==1.20.1
